Integrate trait solver infra to impl matcher
