+++ crates/engine/ty/src/trait_selection/mod.rs
@@ -85,12 +85,48 @@ where
     ) -> Result<ExpectedUnique<TraitSelection>, I::Error> {
         let mut selections = ExpectedUnique::new();
         for trait_impl in self.trait_impl_candidates(goal.trait_ref)? {
-            let Some(selection) = self.probe_impl(goal, table, trait_impl)? else {
+            let Some(selection) = self.probe_trait_impl(goal, table, trait_impl)? else {
                 continue;
             };
             selections.push(selection);
         }
         Ok(selections)
+    }
+
+    /// Probe one already-visible impl against a trait goal.
+    ///
+    /// Method lookup and completion often start from an impl list that was already filtered by
+    /// visibility, receiver indexes, or body-local overlay rules. This entry point lets those
+    /// callers reuse the same bounded header matcher without asking trait selection to enumerate
+    /// candidates again.
+    pub fn probe_trait_impl(
+        &self,
+        goal: &TraitGoal,
+        table: &InferenceTable,
+        trait_impl: TraitImplRef,
+    ) -> Result<Option<TraitSelection>, I::Error> {
+        Self::probe_visible_trait_impl(
+            &self.item_paths,
+            &self.target_items,
+            goal,
+            table,
+            trait_impl,
+        )
+    }
+
+    /// Probe one already-visible impl without constructing an owning query wrapper.
+    ///
+    /// Some callers, such as method lookup, already own borrowed query state and want to reuse the
+    /// same candidate matcher for a single impl. This keeps that integration read-only and avoids
+    /// adding clone requirements to callers that only need a local proof.
+    pub fn probe_visible_trait_impl(
+        item_paths: &ItemPathQuery<'query, D, I>,
+        target_items: &TargetItemQuery<'query, D, I>,
+        goal: &TraitGoal,
+        table: &InferenceTable,
+        trait_impl: TraitImplRef,
+    ) -> Result<Option<TraitSelection>, I::Error> {
+        Self::probe_impl(item_paths, target_items, goal, table, trait_impl)
     }
 
     fn trait_impl_candidates(&self, trait_ref: TraitRef) -> Result<Vec<TraitImplRef>, I::Error> {
@@ -108,12 +144,13 @@ where
     }
 
     fn probe_impl(
-        &self,
+        item_paths: &ItemPathQuery<'query, D, I>,
+        target_items: &TargetItemQuery<'query, D, I>,
         goal: &TraitGoal,
         table: &InferenceTable,
         trait_impl: TraitImplRef,
     ) -> Result<Option<TraitSelection>, I::Error> {
-        let Some(impl_data) = self.target_items.items().impl_data(trait_impl.impl_ref)? else {
+        let Some(impl_data) = target_items.items().impl_data(trait_impl.impl_ref)? else {
             return Ok(None);
         };
         if !impl_data.resolved_trait_ref.is(&goal.trait_ref)
@@ -124,7 +161,7 @@ where
 
         let mut table = table.clone();
         let mut subst = InferTypeSubst::new();
-        let matcher = CandidateMatcher::new(&self.item_paths);
+        let matcher = CandidateMatcher::new(item_paths);
         let Some(applicability) =
             matcher.match_goal(goal, trait_impl, impl_data, &mut table, &mut subst)?
         else {
